Code cleaning in net pkg
- uniformize the file name convention
- code cleaning in net_linux.go
- move env variables to env.go file
